+++ Nimesuki_Api_Rest/src/main/java/proyecto/nimesuki/modelo/Favoritos.java
@@ -32,12 +32,12 @@ public class Favoritos implements Serializable {
     private Usuario usuario;
 
     @Column
-    private Integer valoracion;
+    private Float valoracion;
 
     @Column(name = "cap_actual")
     private Integer capActual;
 
-    public Favoritos(FavoritosId id, Anime anime, Usuario usuario, Integer valoracion, Integer capActual) {
+    public Favoritos(FavoritosId id, Anime anime, Usuario usuario, Float valoracion, Integer capActual) {
         this.id = id;
         this.anime = anime;
         this.usuario = usuario;
@@ -76,11 +76,11 @@ public class Favoritos implements Serializable {
         this.usuario = usuario;
     }
 
-    public Integer getValoracion() {
+    public Float getValoracion() {
         return valoracion;
     }
 
-    public void setValoracion(Integer valoracion) {
+    public void setValoracion(Float valoracion) {
         this.valoracion = valoracion;
     }
